+++ src/tools/style/index.ts
@@ -1,6 +1,6 @@
 import type { FigmaApiClient } from '../../api/figma-api-client.js';
 import { createGetStylesTool } from './list.js';
-import type { StyleTool } from './list.js';
+import type { StyleTool } from './types.js';
 
 interface StyleTools {
   getStyles: StyleTool;
